Refactor all functions out of resourceLoaders.ts for easier mocking. New test over ResourceLoader.getTopicsForCluster().
